Reorganize tests for clarity
